% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupNorm.R
\name{Normalidad}
\alias{Normalidad}
\title{SDENORM}
\usage{
Normalidad(x, alpha)
}
\arguments{
\item{x}{(vector) datos de la muestra.}

\item{alpha}{(vector) nivel de significancia para el test de Shapiro-Wilk}
}
\value{
Un grafico Q-Q para visualizar los datos y un True o False si se cumple la normalidad.
}
\description{
Realiza una prueba de normalidad.
}
